[backend/frontend] Clean
