+++ openbas-api/src/main/java/io/openbas/utils/EndpointMapper.java
@@ -1,32 +1,34 @@
 package io.openbas.utils;
 
 import static io.openbas.database.model.Endpoint.*;
-import static io.openbas.utils.AgentUtils.isPrimaryAgent;
-import static java.util.Collections.emptyList;
+import static io.openbas.utils.AgentUtils.getPrimaryAgents;
 import static java.util.Collections.emptySet;
 
-import io.openbas.database.model.Agent;
 import io.openbas.database.model.Asset;
 import io.openbas.database.model.Endpoint;
 import io.openbas.database.model.Tag;
-import io.openbas.rest.asset.endpoint.form.*;
-import java.util.*;
+import io.openbas.rest.asset.endpoint.form.EndpointOutput;
+import io.openbas.rest.asset.endpoint.form.EndpointOverviewOutput;
+import io.openbas.rest.asset.endpoint.form.EndpointSimple;
+import java.util.Arrays;
+import java.util.HashSet;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 import lombok.RequiredArgsConstructor;
-import org.jetbrains.annotations.NotNull;
 import org.springframework.stereotype.Component;
 
 @Component
 @RequiredArgsConstructor
 public class EndpointMapper {
 
+  final AgentMapper agentMapper;
+
   public EndpointOutput toEndpointOutput(Endpoint endpoint) {
     return EndpointOutput.builder()
         .id(endpoint.getId())
         .name(endpoint.getName())
         .type(endpoint.getType())
-        .agents(toAgentOutputs(getPrimaryAgents(endpoint)))
+        .agents(agentMapper.toAgentOutputs(getPrimaryAgents(endpoint)))
         .platform(endpoint.getPlatform())
         .arch(endpoint.getArch())
         .tags(endpoint.getTags().stream().map(Tag::getId).collect(Collectors.toSet()))
@@ -54,43 +56,9 @@ public class EndpointMapper {
             endpoint.getMacAddresses() != null
                 ? new HashSet<>(Arrays.asList(endpoint.getMacAddresses()))
                 : emptySet())
-        .agents(toAgentOutputs(getPrimaryAgents(endpoint)))
+        .agents(agentMapper.toAgentOutputs(getPrimaryAgents(endpoint)))
         .tags(endpoint.getTags().stream().map(Tag::getId).collect(Collectors.toSet()))
         .build();
-  }
-
-  private Set<AgentOutput> toAgentOutputs(List<Agent> agents) {
-    return Optional.ofNullable(agents).orElse(emptyList()).stream()
-        .map(this::toAgentOutput)
-        .collect(Collectors.toSet());
-  }
-
-  public AgentOutput toAgentOutput(Agent agent) {
-    AgentOutput.AgentOutputBuilder builder =
-        AgentOutput.builder()
-            .id(agent.getId())
-            .privilege(agent.getPrivilege())
-            .deploymentMode(agent.getDeploymentMode())
-            .executedByUser(agent.getExecutedByUser())
-            .isActive(agent.isActive())
-            .lastSeen(agent.getLastSeen());
-
-    if (agent.getExecutor() != null) {
-      builder.executor(
-          ExecutorOutput.builder()
-              .id(agent.getExecutor().getId())
-              .name(agent.getExecutor().getName())
-              .type(agent.getExecutor().getType())
-              .build());
-    }
-    return builder.build();
-  }
-
-  @NotNull
-  private static List<Agent> getPrimaryAgents(Endpoint endpoint) {
-    return endpoint.getAgents().stream()
-        .filter(agent -> isPrimaryAgent(agent))
-        .collect(Collectors.toList());
   }
 
   public static String[] setMacAddresses(String[] macAddresses) {
